requirements.txt
Flask
requests
python-dotenv
pycryptodome
flask_httpauth
Fernet